Intermediate state on way to replacement of local host configuration instructions.
I'm working out and trying to document the last remaining details. At the moment, Firefox is launching rather than Chrome, and it appears that only Chrome has the extension needed to present the debug button that you the user must depress in order to finally launch the app on the emulated phone. I know Amanda solved this one. I think I changed to Chrome as my default browser, but now I'll reboot and try again. Stay tuned to this channel.